[Bronze II] Title: 숫자, Time: 1208 ms, Memory: 28388 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze II/10093. 숫자/숫자.java
@@ -7,10 +7,10 @@ public class Main {
     public static void main(String[] args) throws IOException{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int n1 = Integer.parseInt(st.nextToken());
-        int n2 = Integer.parseInt(st.nextToken());
-        
-        int a = 0, b = 0;
+        long n1 = Long.parseLong(st.nextToken());
+        long n2 = Long.parseLong(st.nextToken());
+
+        long a = 0, b = 0;
         if(n1 <= n2){
             a = n1;
             b = n2;
@@ -19,10 +19,10 @@ public class Main {
             a = n2;
             b = n1;
         }
-        
+
         if(a!=b){
             System.out.println(b-a-1);
-            for(int i=a+1; i<b; i++)
+            for(long i=a+1; i<b; i++)
                 System.out.print(i + " ");
         }
         else System.out.println(0);
